UI,shared: Move OBSPlainTextEdit to its own directory
