+++ UI/frontend-plugins/decklink-output-ui/CMakeLists.txt
@@ -20,6 +20,10 @@ endif()
 add_library(decklink-output-ui MODULE)
 add_library(OBS::decklink-output-ui ALIAS decklink-output-ui)
 
+if(NOT TARGET OBS::qt-plain-text-edit)
+  add_subdirectory("${CMAKE_SOURCE_DIR}/shared/qt/plain-text-edit" "${CMAKE_BINARY_DIR}/shared/qt/plain-text-edit")
+endif()
+
 if(NOT TARGET OBS::qt-wrappers)
   add_subdirectory("${CMAKE_SOURCE_DIR}/shared/qt/wrappers" "${CMAKE_BINARY_DIR}/shared/qt/wrappers")
 endif()
@@ -35,6 +39,7 @@ target_link_libraries(
   decklink-output-ui
   PRIVATE OBS::libobs
           OBS::frontend-api
+          OBS::qt-plain-text-edit
           OBS::qt-wrappers
           OBS::ui-support
           Qt::Widgets
