+++ UI/frontend-plugins/frontend-tools/scripts.cpp
@@ -1,7 +1,6 @@
 #include "obs-module.h"
 #include "scripts.hpp"
 #include "../../properties-view.hpp"
-#include "../../plain-text-edit.hpp"
 
 #include <QFileDialog>
 #include <QHBoxLayout>
@@ -18,6 +17,7 @@
 #include <QUrl>
 #include <QDesktopServices>
 #include <qt-wrappers.hpp>
+#include <plain-text-edit.hpp>
 
 #include <obs.hpp>
 #include <obs-module.h>
