+++ UI/properties-view.cpp
@@ -29,10 +29,10 @@
 #include "spinbox-ignorewheel.hpp"
 #include "properties-view.hpp"
 #include "properties-view.moc.hpp"
-#include "plain-text-edit.hpp"
 #include "obs-app.hpp"
 
 #include <qt-wrappers.hpp>
+#include <plain-text-edit.hpp>
 #include <cstdlib>
 #include <initializer_list>
 #include <obs-data.h>
