+++ shared/qt/plain-text-edit/plain-text-edit.hpp
